+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Dialog Insight API refactoring
+
 ### 1.0.1: March 1st, 2021
 * Encode token in URL
 * Update forms CSS styling